+++ modules/toolRecognizer/include/toolRecognizer_IDLServer.h
@@ -27,12 +27,12 @@ public:
    * Command to train tools by their label
    * @return true/false on success/failure to train classifiers.
    */
-  virtual bool train(const std::string& label, const double tlx = 0, const double tly = 0, const double brx = 0, const double bry = 0);
+  virtual bool train(const std::string& label, const int32_t tlx = 0, const int32_t tly = 0, const int32_t brx = 0, const int32_t bry = 0);
   /**
    * Checks whether the hand is full or empty
    * @return true/false  corresponding to full or empty hand
    */
-  virtual std::string recognize(const double tlx = 0, const double tly = 0, const double brx = 0, const double bry = 0);
+  virtual std::string recognize(const int32_t tlx = 0, const int32_t tly = 0, const int32_t brx = 0, const int32_t bry = 0);
   virtual bool read(yarp::os::ConnectionReader& connection);
   virtual std::vector<std::string> help(const std::string& functionName="--all");
 };
